requirements.txt
# requirements.txt
# Core dependencies
hydra-core>=1.3.0
omegaconf>=2.3.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Machine Learning
torch>=2.0.0
transformers>=4.21.0
xgboost>=1.7.0
lightgbm>=4.0.0  # Added for uncertainty calibration

# API and HTTP
requests>=2.28.0
aiohttp>=3.8.0

# Data processing
pyyaml>=6.0
jsonlines>=3.1.0

# Evaluation and metrics
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Networking and graph analysis
networkx>=3.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Jupyter and notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# Logging and monitoring
wandb>=0.13.0
tensorboard>=2.10.0

# Hyperparameter optimization (optional)
optuna>=3.0.0

# Model serialization
joblib>=1.2.0

# Environment management
python-dotenv>=0.19.0

# Type checking
types-requests>=2.28.0
types-PyYAML>=6.0.0
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Additional API dependencies
python-multipart>=0.0.6  # For form data support if needed